+++ crates/jsonschema-py/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.38.0] - 2025-12-24
+
 ### Added
 
 - `EmailOptions` for configuring `email` format validation.
@@ -807,7 +809,8 @@
 ## 0.1.0 - 2020-06-09
 - Initial public release
 
-[Unreleased]: https://github.com/Stranger6667/jsonschema/compare/python-v0.37.4...HEAD
+[Unreleased]: https://github.com/Stranger6667/jsonschema/compare/python-v0.38.0...HEAD
+[0.38.0]: https://github.com/Stranger6667/jsonschema/compare/python-v0.37.4...python-v0.38.0
 [0.37.4]: https://github.com/Stranger6667/jsonschema/compare/python-v0.37.3...python-v0.37.4
 [0.37.3]: https://github.com/Stranger6667/jsonschema/compare/python-v0.37.2...python-v0.37.3
 [0.37.2]: https://github.com/Stranger6667/jsonschema/compare/python-v0.37.1...python-v0.37.2